requirements.txt
streamlit
langchain
langchain-google-genai>=0.0.30
google-generativeai>=0.7.0
faiss-cpu
python-dotenv
grpcio>=1.63.0
grpcio-tools
langchain_community